feat: Major version upgrade to 2.0.0 with UI overhaul and architecture simplification
- Bump version to 2.0.0 in manifest.json
- Redesign popup.html for a cleaner interface, including a new header and footer
- Remove deprecated background service worker and legacy configuration files
- Introduce a new stylesheet with updated design tokens for improved styling
- Archive previous version styles and scripts for future reference
- Prepare for future development with a modular approach and clean slate for codebase
