+++ src/manifest.json
@@ -1,7 +1,7 @@
 {
   "manifest_version": 3,
   "name": "MarkMind",
-  "version": "1.0.1",
+  "version": "2.0.0",
   "description": "AI-powered bookmark organization that helps you keep your bookmarks clean and organized using artificial intelligence.",
   "permissions": [
     "bookmarks",
@@ -10,9 +10,6 @@
   ],
   "action": {
     "default_popup": "popup.html"
-  },
-  "background": {
-    "service_worker": "js/background.js"
   },
   "icons": {
     "16": "assets/icons/icon16.png",
