+++ src/v1/assets/icons/github.png
@@ -1,0 +1,1 @@
+404: Not Found
