+++ src/v1/docs/PERMISSIONS.md
