+++ src/v1/js/components/LogPanel.js
