+++ src/v1/js/components/ProgressBar.js
