+++ src/v1/js/components/ResultsPreview.js
