+++ src/v1/js/managers/SettingsManager.js
