+++ src/v1/js/managers/UIManager.js
