+++ src/v1/js/services/geminiService.js
